requirements.txt
streamlit==1.32.2
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
numpy==1.26.4
